+++ database/migrations/2026_04_02_000108_create_tester_calibration_schedules_table.php
@@ -17,11 +17,11 @@ return new class extends Migration
             $table->dateTime('last_calibration_date')->nullable();
             $table->dateTime('next_calibration_due')->nullable();
 
-            $table->unsignedInteger('tester_id');
-            $table->unsignedInteger('calibration_id');
-            $table->unsignedInteger('calibration_status')->nullable();
-            $table->unsignedInteger('last_calibration_by_user_id')->nullable();
-            $table->unsignedInteger('next_calibration_by_user_id')->nullable();
+            $table->unsignedBigInteger('tester_id');
+            $table->unsignedBigInteger('calibration_id');
+            $table->unsignedBigInteger('calibration_status')->nullable();
+            $table->unsignedBigInteger('last_calibration_by_user_id')->nullable();
+            $table->unsignedBigInteger('next_calibration_by_user_id')->nullable();
 
             // Index for faster lookups
             $table->index('tester_id', 'idx_tester_calibration_schedules_tester');
